Add mypy type checking, Hypothesis property tests, pre-commit hooks, and coverage improvements
- mypy passes clean across all source files (0 errors)
- Added mypy config to pyproject.toml with per-module overrides
- Added mypy CI job to GitHub Actions workflow
- Fixed 5 type issues: np.nan returns, TypedDict dynamic keys, matplotlib Figure types
- Added Hypothesis property-based tests (14 tests) verifying numerical invariants:
  probability sums, stochastic matrices, feature vector round-trips, pipeline output
- Added 6 tests for fetch_multi_asset covering ref columns, caret stripping,
  error warnings, empty data, missing columns
- Added pre-commit config with ruff, trailing whitespace, and merge conflict checks
- Updated dev dependencies: hypothesis, mypy, types-PyYAML
- Added `make typecheck` target and .mypy_cache to clean
- All ruff formatting applied consistently
- 260 tests passing, 98% coverage

+++ src/financial_dynamics/forecasting/__init__.py
@@ -19,6 +19,7 @@ __all__ = [
 @dataclass
 class RegimeForecast:
     """K-step-ahead regime probability forecast."""
+
     current_regime: Regime
     current_probabilities: RegimeProbabilities
     horizon_probabilities: list[RegimeProbabilities]
@@ -81,7 +82,7 @@ def compute_expected_duration(
     self_transition = transition_matrix[int(regime), int(regime)]
     if self_transition >= 1.0:
         return float("inf")
-    return 1.0 / (1.0 - self_transition)
+    return float(1.0 / (1.0 - self_transition))
 
 
 def compute_stationary_distribution(
